% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_bed_file_paths}
\alias{get_bed_file_paths}
\title{get_bed_file_paths}
\usage{
get_bed_file_paths()
}
\value{
(chr) vector of input xlsx file paths.
}
\description{
Get all the paths to the xlsx files containing bed data and classes
to use for classification.
}
\note{
Input data path is automatically retrieved from the
corresponding entry in the project \code{.Renviron} files (you can open
it by running \code{usethis::edit_r_environ("project")})
}
\examples{
get_bed_file_paths()
}
