% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{import_bed}
\alias{import_bed}
\title{import_bed}
\usage{
import_bed(bed_path)
}
\arguments{
\item{bed_path}{(chr) path to a (single) xlsx file of bed date.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Read the xlsx files from disk to a \link[tibble:tibble-package]{tibble}.
}
\examples{
if (FALSE) {
  import_bed("path/to/bed_data.xlsx")
}
}
