% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iobed.model-package.R
\docType{package}
\name{iobed.model-package}
\alias{iobed.model}
\alias{iobed.model-package}
\title{iobed.model: DL models for I.O.BED project at UBEP}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provide functions, utilities and models to classify person movement on a smart bed basted on weight sensors under the bed.
}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@ubep.unipd.it} (\href{https://orcid.org/0000-0002-0520-7428}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Alessandra Casamento \email{alessandra.casamento@ubep.unipd.it}
}

Other contributors:
\itemize{
  \item Silvano Salaris \email{silvano.salaris@ubep.unipd.it} (\href{https://orcid.org/0000-0002-8389-7460}{ORCID}) [contributor]
  \item Dario Gregori \email{dario.gregori@ubep.unipd.it} (\href{https://orcid.org/0000-0001-7906-0580}{ORCID}) [contributor, copyright holder]
}

}
\keyword{internal}
