% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepare_supervised_bed}
\alias{prepare_supervised_bed}
\title{prepare_supervised_bed}
\usage{
prepare_supervised_bed(db, label_dict, include_casual = FALSE)
}
\arguments{
\item{db}{(data.frame) bed raw data including the sensors information
and the labels for by-row classification.}

\item{include_casual}{(lgl, default FALSE) include starting "casual"
movement?}
}
\value{
list of two array named "x" and "y_true" containing
respectively the sensors/input data and the true responses for the
classification
}
\description{
Data from a single IOBED experiments included four weighting sensors
reporting the overall weight lying on the bed and the percentage of
weight distribution across each of them. In particular thy can be:
}
\details{
\itemize{
\item \strong{sbl} (Sensor Bottom Left): \link{0-1000}
\item \strong{sul} (Sensor Upper Left): \link{0-1000}
\item \strong{sbr} (Sensor Bottom Right): \link{0-1000}
\item \strong{sur} (Sensor Upper Right): \link{0-1000}
\item \strong{weight} (integer hg): e.g. 800 are 80.0 kg
}

Moreover information of the two possible positions of the bed
backrest tilt (\code{tilt_bed}) of 0 or 30 degrees (encoded as 0/1
correspondingly).

Classification outcome is composed by four pieces of information,
i.e.: position (\code{static}) and dynamics (\code{dyn}) of the subject wrt the
bed (\code{bed}) or themself (\code{self}). In particular there are the
following classes for each category:
\itemize{
\item \strong{static_bed}: Center (2), Right (1), Left (3), Transition (-1),
Null (0)
\item \strong{static_self}: Supine (2), Right (1), Left (3), Transition (-1),
Null (0)
\item \strong{dyn_bed}: Static (2), Slide to right (1), Slide to left (3),
Entrance Right (4), Entrance Left (5), Exit Right (6),
Exit Left (7), Null (0)
\item \strong{dyn_self}: Static (2), Turn to right (1), Turn to left (3),
Entrance Right (4), Entrance Left (5), Exit Right (6),
Exit Left (7), Assessment (8), Null (0)
}
}
\note{
Time is not reported inside the data because it is supposed
to flow at approximately 300 ms (+- 2 ms)  each row.

During the experiment people was asked to move randomly on the
bed. in a first phase of the modeling we ignore that portion of
data entirely labelled with \code{casual}.
}
\examples{
if (FALSE) {
 db <- import_bed("path/to/bed_data.xlsx")
 db_train <- prepare_supervised_bed(bed)
}
}
